tiny_multi_head_lut: add tests; fix hybrid_smooth output dtype
Add src/spiky/lutorch/tests/test_tiny_multi_head_lut.py (33 tests):
 - Forward shape / dtype across forward_mode x weight_dtype.
 - Forward determinism, eval == train forward, runtime forward_mode flip.
 - Argument validation (bad forward_mode, NAP out of range, unsupported
   anchor sampling policy).
 - Anchor sampling invariants (a != b within table; same seed -> same
   pairs; different seeds -> different).
 - Backward grad flow on x, weights, learnable log temperatures.
 - weights.grad dtype matches weights.dtype (the autograd same-dtype
   invariant for fp32 master weights).
 - Frozen temps are buffers with no grad.
 - hybrid_smooth -> hard collapse for high-confidence inputs (u -> 0).
 - Soft surrogate backward agrees with a finite-difference reference of
   the full-soft forward (direction + magnitude).
 - Smoke training step under bf16 autocast produces no NaN/inf.

Fix that the tests caught: _hybrid_smooth_lut_fwd_body returned fp32 even
when weight_dtype was bf16, because bf16 (rows) * fp32 (1-u) promotes to
fp32 and the trailing .sum() is autocast-promoted to fp32. Cast the
final output to weights.dtype so out.dtype == weights.dtype in both
forward modes, matching the docstring.

+++ src/spiky/lutorch/tiny_multi_head_lut.py
@@ -299,7 +299,10 @@ def _hybrid_smooth_lut_fwd_body(x, weights, anchor_a_long, anchor_b_long, powers
     main_rows = F.embedding(main_flat_idx, weights_flat).view(B, n_tables, n_outputs)
     alt_rows  = F.embedding(alt_flat_idx,  weights_flat).view(B, n_tables, n_outputs)
     blended = main_rows * main_w.unsqueeze(-1) + alt_rows * u.unsqueeze(-1)
-    out = blended.view(B, n_heads, tph, n_outputs).sum(dim=2)
+    # Match the hard path's contract that out.dtype == weights.dtype. Under
+    # bf16 autocast .sum() is promoted to fp32 for stability, so an explicit
+    # final cast is needed.
+    out = blended.view(B, n_heads, tph, n_outputs).sum(dim=2).to(weights.dtype)
     return out, main_index, alt_index, u
